requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.0
build==0.10.0
certifi @ file:///C:/ci_311/certifi_1676421629969/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
docopt==0.6.2
exceptiongroup==1.1.1
gTTS==2.3.1
h11==0.14.0
idna==3.4
MouseInfo==0.1.3
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
pip-tools==6.12.3
pipreqs==0.4.11
playsound==1.3.0
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
pyproject_hooks==1.0.0
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dotenv==1.0.0
pytweening==1.0.4
requests==2.28.2
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SpeechRecognition==3.10.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
webdriver-manager==3.8.5
wikipedia==1.4.0
wincertstore==0.2
wsproto==1.2.0
yarg==0.1.9
youtube-search==2.1.2
